Avoid coalescing endpoint regeneration if regeneration level may differ
The RegenerateIfAlive function seeks to skip an endpoint regeneration if
there is already a regeneration pending for that endpoint (i.e. it is in
StateRestoring or StateWaitingToRegenerate).   This makes sense if all
regenerations are equal.  However, we also introduced the notion of a
requested "Regeneration Level", so that each regeneration is only as
"deep" (and therefore expensive) as is needed by the function starting
the regeneration.   For example, discovering a new identity in the system
does not require fully reloading the BPF datapath, so it calls regenerate
with level RegenerateWithoutDatapath, which is cheaper than a call with
RegenerateWithDatapathLoad, and so forth.

However, a negative interaction can happen here where a pending regeneration
for a lower regeneration level (e.g., RegenerateWithoutDatapath) results
in us skipping a regeneration with a higher regeneration level (e.g.,
RegenerateWithDatapathLoad).   Since the regeneration level of the already
queued regeneration is not immediately available in the RegenerateIfAlive
function, this patch takes a simple approach of only skipping a regeneration
if it is the lowest regeneration level (i.e., RegenerateWithoutDatapath) as
any other regeneration must be at least that deep.   While not ideal, in
practice the most common type of regeneration is RegenerateWithDatapathLoad
(e.g., when new identities are discovered in the system), so we still get
most of the optimization, but now have better correctness guarantees as well.

This issue was identified because we found a race condition that resulted in
a missing ipcache entries.  The cause is that the ipcache GC on 4.11-4.15 kernels
requires a RegenerateWithDatapathLoad regeneration in order to swap in a
new ipcache map every GC interval (5 minutes).   There was a race that if a
another regeneration that was less deep was happening at the same time, the
RegenerateWithDatapathLoad load generation would be skipped, leading to a
stale ipcache until the GC would run again.   In this environment, there was
a policy with "toEntities: cluster" and new identities being created on
a frequent basis, meaning that each time a new identity was created, all
endpoints in the system would have a RegenerateWithoutDatapath regeneration.
If one of those less deep regenerations happened right before a GC regeneration
for the same endpoint, the deeper RegenerateWithDatapathLoad would not happen,
leading to the use of an outdated ipcache until the next GC.

Signed-off-by: Dan Wendlandt <dan@covalent.io>